server/requirements.txt
openenv-core>=0.1.0
stable-baselines3[extra]>=2.3.0
gymnasium>=0.29.0
torch>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
requests>=2.31.0
